+++ src/pretalx/static/common/js/formTools.js
@@ -1,3 +1,4 @@
+/* Handle Markdown */
 let dirtyInputs = []
 const options = {
   baseUrl: null,
@@ -29,6 +30,22 @@ function checkForChanges() {
   }
   checkChangeTimeout = window.setTimeout(checkForChanges, 100)
 }
+
+const warnFileSize = (element) => {
+  const warning = document.createElement("div")
+  warning.classList = ["invalid-feedback"]
+  console.log(warning)
+  warning.textContent = element.dataset.sizewarning
+  element.parentElement.appendChild(warning)
+  element.classList.add("is-invalid")
+}
+const unwarnFileSize = (element) => {
+  element.classList.remove("is-invalid")
+  const warning = element.parentElement.querySelector(".invalid-feedback")
+  if (warning) element.parentElement.removeChild(warning)
+}
+
+/* Register handlers */
 window.onload = () => {
   document.querySelectorAll(".markdown-wrapper").forEach(element => {
     const inputElement = element.querySelector("textarea")
@@ -43,4 +60,17 @@ window.onload = () => {
     inputElement.addEventListener("keydown", handleInput, false)
   })
   checkForChanges()
+
+  document.querySelectorAll("input[data-maxsize][type=file]").forEach(element => {
+    const checkFileSize = () => {
+      const files = element.files
+      if (!files || !files.length) {
+        unwarnFileSize(element)
+      } else {
+        maxsize = parseInt(element.dataset.maxsize)
+        if (files[0].size > maxsize) { warnFileSize(element) } else { unwarnFileSize(element) }
+      }
+    }
+    element.addEventListener("change", checkFileSize, false)
+  })
 }